refactor: rename ContentHero to PageHero
