+++ app/[locale]/community/support/page.tsx
@@ -4,7 +4,7 @@ import { getTranslations, setRequestLocale } from "next-intl/server"
 import type { Lang, PageParams } from "@/lib/types"
 
 import FeedbackCard from "@/components/FeedbackCard"
-import ContentHero from "@/components/Hero/ContentHero"
+import PageHero from "@/components/Hero/PageHero"
 import MainArticle from "@/components/MainArticle"
 import {
   Alert,
@@ -44,7 +44,7 @@ export default async function Page(props: { params: Promise<PageParams> }) {
     <div>
       <SupportJsonLD locale={locale} contributors={contributors} />
       {/* Hero */}
-      <ContentHero
+      <PageHero
         breadcrumbs={{ slug: "/community/support", startDepth: 1 }}
         title={t("page-community-support-hero-title")}
         description={
