+++ app/[locale]/developers/tutorials/page.tsx
@@ -8,7 +8,7 @@ import {
 import type { Lang, PageParams } from "@/lib/types"
 
 import FeedbackCard from "@/components/FeedbackCard"
-import ContentHero from "@/components/Hero/ContentHero"
+import PageHero from "@/components/Hero/PageHero"
 import I18nProvider from "@/components/I18nProvider"
 import MainArticle from "@/components/MainArticle"
 
@@ -57,7 +57,7 @@ const Page = async (props: { params: Promise<PageParams> }) => {
         contributors={contributors}
       />
       <I18nProvider locale={locale} messages={messages}>
-        <ContentHero
+        <PageHero
           breadcrumbs={{ slug: "developers/tutorials", startDepth: 1 }}
           heroImg={heroImg}
           title={t("page-tutorial-title")}
