+++ app/[locale]/ethereum-vs-bitcoin/page.tsx
@@ -3,7 +3,7 @@ import { getTranslations, setRequestLocale } from "next-intl/server"
 import type { Lang, ToCItem } from "@/lib/types"
 
 import FileContributors from "@/components/FileContributors"
-import ContentHero from "@/components/Hero/ContentHero"
+import PageHero from "@/components/Hero/PageHero"
 import { Image } from "@/components/Image"
 import { Strong } from "@/components/IntlStringElements"
 import MainArticle from "@/components/MainArticle"
@@ -67,7 +67,7 @@ const Page = async ({ params }: { params: Promise<{ locale: Lang }> }) => {
         contributors={contributors}
       />
 
-      <ContentHero
+      <PageHero
         breadcrumbs={{ slug: "learn/ethereum-vs-bitcoin", startDepth: 1 }}
         heroImg={heroImg}
         title={t("page-ethereum-vs-bitcoin-title")}
