+++ app/[locale]/layer-2/networks/page.tsx
@@ -7,7 +7,7 @@ import {
 
 import type { ExtendedRollup, Lang, PageParams } from "@/lib/types"
 
-import { ContentHero } from "@/components/Hero"
+import PageHero from "@/components/Hero/PageHero"
 import I18nProvider from "@/components/I18nProvider"
 import Layer2NetworksTable from "@/components/Layer2NetworksTable"
 import MainArticle from "@/components/MainArticle"
@@ -155,7 +155,7 @@ const Page = async (props: { params: Promise<PageParams> }) => {
         contributors={contributors}
       />
       <MainArticle className="relative flex flex-col">
-        <ContentHero
+        <PageHero
           breadcrumbs={{ slug: "/layer-2/networks", startDepth: 1 }}
           heroImg={heroImg}
           title={tCommon("nav-networks-explore-networks-label")}
