+++ src/components/Hero/index.ts
@@ -1,7 +1,7 @@
-export { default as ContentHero, type ContentHeroProps } from "./ContentHero"
 export { default as HomeHero } from "./HomeHero"
 export { default as HubHero } from "./HubHero"
 export { default as MdxHero, type MdxHeroProps } from "./MdxHero"
+export { default as PageHero, type PageHeroProps } from "./PageHero"
 export {
   default as SimpleHero,
   type SimpleHeroButton,
